requirements.txt
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
et-xmlfile==1.1.0
idna==3.4
numpy==1.23.4
openpyxl==3.0.10
pandas==1.5.1
PyAudio==0.2.12
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.5
requests==2.28.1
six==1.16.0
srt==3.5.2
tqdm==4.64.1
urllib3==1.26.12
vosk==0.3.44
websockets==10.4
